refactor(agents): decompose AgentEditorScreen into sibling files
Split the 926-line AgentEditorScreen into a 141-line shell plus four
focused sibling files in the same package: AgentEditorTopBar,
AgentEditorDialogs, AgentEditorComponents (InstructionsField,
SelectedToolRow), and AgentEditorForm. The screen retains state
wiring, effects, and the Scaffold composing them. Pure refactor;
public signature and both navigation call sites unchanged.
